changed the entire receipt systemand also added collection planner.

+++ src/main/java/com/rt/purchases/PurchaseInvoiceItem.java
@@ -21,23 +21,23 @@ public class PurchaseInvoiceItem {
     @JoinColumn(name = "product_id")
     private Product product;
 
-    private Integer quantity;
+    // --- CHANGED TO DOUBLE ---
+    private Double quantity; 
     private Double purchasePrice;
 
     @ManyToOne
     @JoinColumn(name = "purchase_invoice_id")
-    @JsonIgnore
+    @JsonIgnore 
     private PurchaseInvoice purchaseInvoice;
 
-    // Getters and Setters
     public Long getId() { return id; }
     public void setId(Long id) { this.id = id; }
 
     public Product getProduct() { return product; }
     public void setProduct(Product product) { this.product = product; }
 
-    public Integer getQuantity() { return quantity; }
-    public void setQuantity(Integer quantity) { this.quantity = quantity; }
+    public Double getQuantity() { return quantity; }
+    public void setQuantity(Double quantity) { this.quantity = quantity; }
 
     public Double getPurchasePrice() { return purchasePrice; }
     public void setPurchasePrice(Double purchasePrice) { this.purchasePrice = purchasePrice; }
